+++ NEWS.md
@@ -1,6 +1,10 @@
-csquares v0.0.3.005 (Release date: TODO)
+csquares v0.0.4 (Release date: 2024-07-05)
 -------------
 
+  * Added function to expand wildcards
+  * Added function to match wildcards
+  * Added vignette about wildcards
+  * Added multiple methods for base and tidyverse S3 generics
   * Added self tests
   * Several minor corrections